+++ src/Exception/EnumNotRegisteredException.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 /*
  * This file is part of the Enumer package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://wakeapp.ru>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Component\Enumer\Exception;
+namespace MarfaTech\Component\Enumer\Exception;
 
-class EnumNotRegisteredException extends WakeappEnumerException
+class EnumNotRegisteredException extends MarfatechEnumerException
 {
 }